+++ assets/src/application/ui/forms/espressoForm/TestForm.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { formConfig } from './config';
+import EspressoForm from './EspressoForm';
+
+const TestForm = () => <EspressoForm {...formConfig} />;
+
+export default TestForm;
