+++ assets/src/application/ui/forms/espressoForm/index.tsx
@@ -1,32 +1,5 @@
-import React from 'react';
-import { Form as ReactFinalForm } from 'react-final-form';
-import arrayMutators from 'final-form-arrays';
+export { default as EspressoForm } from './EspressoForm';
 
-import { EspressoFormProps } from './types';
-import FormRenderer from './renderers/FormRenderer';
-import { formConfig } from './config';
-import { FormProvider } from './context';
-import './styles.scss';
+export * from './types';
 
-const EspressoForm: React.FC<EspressoFormProps> = ({ onSubmit, mutators, layout, ...rest }) => {
-	const context = { layout };
-	return (
-		<div className='ee-form'>
-			<FormProvider value={context}>
-				<ReactFinalForm
-					onSubmit={onSubmit}
-					render={FormRenderer}
-					mutators={{
-						...arrayMutators,
-						...mutators,
-					}}
-					{...rest}
-				/>
-			</FormProvider>
-		</div>
-	);
-};
-
-export const TestForm = () => <EspressoForm {...formConfig} />;
-
-export default EspressoForm;
+export { default as TestForm } from './TestForm';
