+++ assets/src/application/ui/forms/espressoForm/types.ts
@@ -17,6 +17,7 @@ interface AdditionalFormProps extends FormContextProps {
 	fields?: FieldList;
 	submitButton?: FormButtonProps;
 	resetButton?: FormButtonProps;
+	formWrapper?: React.ReactType;
 }
 
 export interface AdditionalFieldProps {
